tests/wpt/web-platform-tests/bluetooth/idl/idl-NavigatorBluetooth.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_true('bluetooth' in navigator,
              'navigator.bluetooth exists.');
}, 'navigator.bluetooth IDL test');

test(() => {
  assert_equals(navigator.bluetooth, navigator.bluetooth);
}, '[SameObject] test for navigator.bluetooth');
</script>
